+++ robe-chat-ui/config/webpack.config.dev.js
@@ -7,7 +7,7 @@ const webPackConfig = require("./webpack.config.common.js")("/src", "/build", "_
 
 webPackConfig.cache = true;
 webPackConfig.debug = true;
-webPackConfig.devtool = "eval-source-map";
+webPackConfig.devtool = "source-map";
 webPackConfig.entry = {
     app: [webPackConfig.paths.app]
 };