+++ robe-chat-ui/src/index.js
@@ -1,10 +1,10 @@
 import React from "react";
 import {render} from "react-dom";
 import Switch from "app/Switch";
-import Application from "robe-react-commons/lib/application/Application";// eslint-disable-line import/no-unresolved
+import Application from "robe-react-commons/lib/application/Application";
 
 const app = document.getElementById("app");
 
-Application.setBaseUrlPath("http://127.0.0.1:8082/robe-chat");
+Application.setBaseUrlPath("http://192.168.1.72:8082/robe-chat");
 
 render((<Switch />), app);