+++ robe-chat-ui/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## 1.0.1 Developing
+* added WebSocketStore to robe-react-commons
+* change files libs/card to app/card